requirements.txt
discord.py
git+https://github.com/Nagim123/PinterestBoardParser
g4f